Added a new factory method for Term
This change influenced a bunch of different files!  Previously, a
constructor like Term(Int, Map[Symbol, Int]) created no problems when
saying Term(0, Map()), for example.  However, with the introduction of a
two arg factory method with the same name as the constructor in scope,
Term(0, Map()) became a problem.

The compiler deduced that the Map argument was parameterized with
Symbol and Int, but Map() is parameterized with Nothing and Nothing.
This created problems as all calls to Term(x, Map()) and the like had to
be replaced with Term(x).